start moving some database stuff around
