+++ internal/federation/federatingdb/following.go
@@ -42,7 +42,7 @@ func (f *federatingDB) Following(c context.Context, actorIRI *url.URL) (followin
 	}
 
 	acctFollowing := []gtsmodel.Follow{}
-	if err := f.db.GetFollowingByAccountID(acct.ID, &acctFollowing); err != nil {
+	if err := f.db.GetAccountFollowing(acct.ID, &acctFollowing); err != nil {
 		return nil, fmt.Errorf("FOLLOWING: db error getting following for account id %s: %s", acct.ID, err)
 	}
