+++ internal/gtsmodel/statusfave.go
@@ -28,17 +28,13 @@ type StatusFave struct {
 	CreatedAt time.Time `pg:"type:timestamp,notnull,default:now()"`
 	// id of the account that created ('did') the fave
 	AccountID string `pg:"type:CHAR(26),notnull"`
+	Account *Account `pg:"rel:has-one"`
 	// id the account owning the faved status
 	TargetAccountID string `pg:"type:CHAR(26),notnull"`
+	TargetAccount *Account `pg:"rel:has-one"`
 	// database id of the status that has been 'faved'
 	StatusID string `pg:"type:CHAR(26),notnull"`
+	Status *Status `pg:"rel:has-one"`
 	// ActivityPub URI of this fave
 	URI string `pg:",notnull"`
-
-	// GTSStatus is the status being interacted with. It won't be put or retrieved from the db, it's just for conveniently passing a pointer around.
-	GTSStatus *Status `pg:"-"`
-	// GTSTargetAccount is the account being interacted with. It won't be put or retrieved from the db, it's just for conveniently passing a pointer around.
-	GTSTargetAccount *Account `pg:"-"`
-	// GTSFavingAccount is the account doing the faving. It won't be put or retrieved from the db, it's just for conveniently passing a pointer around.
-	GTSFavingAccount *Account `pg:"-"`
 }
