+++ internal/processing/followrequest.go
@@ -28,7 +28,7 @@ import (
 
 func (p *processor) FollowRequestsGet(auth *oauth.Auth) ([]apimodel.Account, gtserror.WithCode) {
 	frs := []gtsmodel.FollowRequest{}
-	if err := p.db.GetFollowRequestsForAccountID(auth.Account.ID, &frs); err != nil {
+	if err := p.db.GetAccountFollowRequests(auth.Account.ID, &frs); err != nil {
 		if _, ok := err.(db.ErrNoEntries); !ok {
 			return nil, gtserror.NewErrorInternalError(err)
 		}
